Move schema registry/manager stuff to the kafka package
