+++ safir/src/safir/kafka/__init__.py
@@ -1,4 +1,11 @@
-from ._config import (
+from ._exceptions import (
+    IncompatibleSchemaError,
+    InvalidAvroNameError,
+    InvalidMetadataError,
+    UnknownDeserializeError,
+    UnknownSchemaError,
+)
+from ._kafka_config import (
     AIOKafkaParams,
     FastStreamBrokerParams,
     KafkaConnectionSettings,
@@ -9,10 +16,23 @@ from ._config import (
     KafkaSecurityProtocol,
     KafkaSslSettings,
 )
+from ._pydantic_schema_manager import (
+    PydanticSchemaManager,
+    SchemaInfo,
+    SchemaRegistryCompatibility,
+)
+from ._schema_registry_config import (
+    SchemaManagerSettings,
+    SchemaRegistryClientParams,
+    SchemaRegistryConnectionSettings,
+)
 
 __all__ = [
     "AIOKafkaParams",
     "FastStreamBrokerParams",
+    "IncompatibleSchemaError",
+    "InvalidAvroNameError",
+    "InvalidMetadataError",
     "KafkaConnectionSettings",
     "KafkaPlaintextSettings",
     "KafkaPlaintextSettings",
@@ -22,4 +42,15 @@ __all__ = [
     "KafkaSecurityProtocol",
     "KafkaSslSettings",
     "KafkaSslSettings",
+    "PydanticSchemaManager",
+    "SchemaInfo",
+    "SchemaManagerSettings",
+    "SchemaManagerSettings",
+    "SchemaRegistryClientParams",
+    "SchemaRegistryClientParams",
+    "SchemaRegistryCompatibility",
+    "SchemaRegistryConnectionSettings",
+    "SchemaRegistryConnectionSettings",
+    "UnknownDeserializeError",
+    "UnknownSchemaError",
 ]
