+++ safir/src/safir/kafka/_exceptions.py
