+++ safir/src/safir/kafka/_kafka_config.py
