+++ safir/src/safir/kafka/_pydantic_schema_manager.py
@@ -28,9 +28,9 @@ from ._exceptions import (
 )
 
 __all__ = [
-    "SchemaRegistryCompatibility",
     "PydanticSchemaManager",
     "SchemaInfo",
+    "SchemaRegistryCompatibility",
 ]
 
 P = TypeVar("P", bound=AvroBaseModel)
