+++ safir/src/safir/kafka/_schema_registry_config.py
