--- safir/src/safir/schema_manager/__init__.py
@@ -1,31 +1,0 @@
-from ._config import (
-    SchemaManagerSettings,
-    SchemaRegistryClientParams,
-    SchemaRegistryConnectionSettings,
-)
-from ._exceptions import (
-    IncompatibleSchemaError,
-    InvalidAvroNameError,
-    InvalidMetadataError,
-    UnknownDeserializeError,
-    UnknownSchemaError,
-)
-from ._pydantic_schema_manager import (
-    PydanticSchemaManager,
-    SchemaInfo,
-    SchemaRegistryCompatibility,
-)
-
-__all__ = [
-    "IncompatibleSchemaError",
-    "InvalidAvroNameError",
-    "InvalidMetadataError",
-    "PydanticSchemaManager",
-    "SchemaInfo",
-    "SchemaManagerSettings",
-    "SchemaRegistryClientParams",
-    "SchemaRegistryCompatibility",
-    "SchemaRegistryConnectionSettings",
-    "UnknownDeserializeError",
-    "UnknownSchemaError",
-]
